+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Unreleased
 
+- Adds performance setting `InactiveCollectionPeriods` allowing
+  control over how fast records are removed from memory.
+
 ## [1.13.1](https://github.com/lightstep/otel-launcher-go/releases/tag/v1.13.1) - 2022-02-15)
 
 - Adds performance improvements for concurrent use of synchronous